Move GetPidAsync method from PidCache to Cluster.
